pyproject.toml
[project]
name = "federated-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fluke-fl>=0.7.9",
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "numpy>=2.4.1",
    "pandas>=3.0.0",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "ucimlrepo>=0.0.7",
]
